Improvements related to LocalStorage and BrowserPageSession
  -- It includes needful javadoc
